refactor(docx-compare): extract revision markup helpers
Move revision ID allocation and the wrapping primitives needed by tagged publication and text-box safety into a portable module before the legacy in-place implementation is deleted. Temporary re-exports keep the legacy stack behaviorally identical at this checkpoint while keeper modules no longer depend on files scheduled for removal.\n\nRef: #838
