+++ packages/docx-compare/src/baselines/atomizer/inPlaceModifier-shared.ts
@@ -8,6 +8,18 @@
 import type { ComparisonUnitAtom } from '@usejunior/docx-core';
 import { childElements, findChildByTagName } from '@usejunior/docx-core';
 import { parseXml } from '@usejunior/docx-core';
+import {
+  allocateRevisionId,
+  type RevisionIdState,
+} from './revisionMarkup.js';
+
+export {
+  allocateRevisionId,
+  createRevisionIdState,
+  formatDate,
+  seedRevisionIdsFromMarkup,
+} from './revisionMarkup.js';
+export type { RevisionIdState } from './revisionMarkup.js';
 
 export const SYNTHETIC_DOC = parseXml('<root xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main"/>');
 export const W_NS = 'http://schemas.openxmlformats.org/wordprocessingml/2006/main';
@@ -37,93 +49,6 @@ export function attachSourceElementPointers(atoms: ComparisonUnitAtom[]): void {
   }
 }
 
-export interface RevisionIdState {
-  nextId: number;
-  /** Canonical decimal identities already present or allocated in this output. */
-  usedRevisionIds: Set<string>;
-  moveRangeIds: Map<string, { sourceRangeId: number; destRangeId: number }>;
-  /** Exact move-range markers generated by this comparison pass. */
-  generatedMoveRangeMarkers: Set<Element>;
-  /** Revision IDs of w:rPrChange records generated by this comparison pass. */
-  generatedFormatChangeIds: Set<string>;
-  /** Track which run elements have already been wrapped */
-  wrappedRuns: Set<Element>;
-  /**
-   * Source bookmark markers cloned into inserted deleted/moveFrom content.
-   * Prevents duplicate marker emission when a single source run is split into
-   * multiple atoms (word-level atomization).
-   */
-  emittedSourceBookmarkMarkers: Set<Element>;
-  /**
-   * Marker clones emitted for the markers that FOLLOWED a source run, keyed by
-   * that source run. Word-level atomization fragments one source run into many
-   * wrappers inserted one after another; each later fragment moves its run's
-   * trailing clones after itself, so they come to rest after the run's final
-   * fragment — where the source markers sat (issue #643).
-   */
-  trailingSourceBookmarkClones: Map<Element, Element[]>;
-}
-
-/**
- * Create initial revision ID state.
- */
-export function createRevisionIdState(preservedRoots: readonly Element[] = []): RevisionIdState {
-  const state: RevisionIdState = {
-    nextId: 1,
-    usedRevisionIds: new Set(),
-    moveRangeIds: new Map(),
-    generatedMoveRangeMarkers: new Set(),
-    generatedFormatChangeIds: new Set(),
-    wrappedRuns: new Set(),
-    emittedSourceBookmarkMarkers: new Set(),
-    trailingSourceBookmarkClones: new Map(),
-  };
-  for (const root of preservedRoots) seedRevisionIdsFromMarkup(state, root);
-  return state;
-}
-
-function canonicalDecimalId(value: string): string | null {
-  const collapsed = value.trim();
-  if (!/^[+-]?\d+$/.test(collapsed)) return null;
-  try {
-    return BigInt(collapsed).toString();
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Reserve every numeric `w:id` in markup that can survive into the comparison
- * output. Numeric aliases such as `1`, `01`, and `+1` reserve one identity.
- *
- * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.23
- * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.24
- * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.27
- * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.28
- * @see https://github.com/UseJunior/safe-docx/issues/446
- */
-export function seedRevisionIdsFromMarkup(state: RevisionIdState, root: Element): void {
-  const visit = (element: Element): void => {
-    const rawId = element.getAttribute('w:id');
-    if (rawId !== null) {
-      const canonical = canonicalDecimalId(rawId);
-      if (canonical !== null) state.usedRevisionIds.add(canonical);
-    }
-    for (const child of childElements(element)) visit(child);
-  };
-  visit(root);
-}
-
-/**
- * Allocate a new revision ID.
- */
-export function allocateRevisionId(state: RevisionIdState): number {
-  while (state.usedRevisionIds.has(String(state.nextId))) state.nextId++;
-  const id = state.nextId++;
-  state.usedRevisionIds.add(String(id));
-  return id;
-}
-
 /**
  * Get or allocate move range IDs for a move name.
  */
@@ -140,13 +65,6 @@ export function getMoveRangeIds(
     state.moveRangeIds.set(moveName, ids);
   }
   return ids;
-}
-
-/**
- * Format date for OOXML (ISO 8601 without milliseconds).
- */
-export function formatDate(date: Date): string {
-  return date.toISOString().replace(/\.\d{3}Z$/, 'Z');
 }
 
 /**
