+++ packages/docx-compare/src/baselines/atomizer/revisionMarkup.ts
@@ -1,0 +1,215 @@
+import {
+  childElements,
+  findChildByTagName,
+  insertAfterElement,
+  parseXml,
+  wrapElement,
+} from '@usejunior/docx-core';
+
+const SYNTHETIC_DOC = parseXml(
+  '<root xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main"/>',
+);
+
+function createRevisionElement(tag: string, attrs?: Record<string, string>): Element {
+  const element = SYNTHETIC_DOC.createElementNS(
+    'http://schemas.openxmlformats.org/wordprocessingml/2006/main',
+    tag,
+  );
+  if (attrs) {
+    for (const [name, value] of Object.entries(attrs)) element.setAttribute(name, value);
+  }
+  return element;
+}
+
+export interface RevisionIdState {
+  nextId: number;
+  /** Canonical decimal identities already present or allocated in this output. */
+  usedRevisionIds: Set<string>;
+  moveRangeIds: Map<string, { sourceRangeId: number; destRangeId: number }>;
+  /** Exact move-range markers generated by this comparison pass. */
+  generatedMoveRangeMarkers: Set<Element>;
+  /** Revision IDs of w:rPrChange records generated by this comparison pass. */
+  generatedFormatChangeIds: Set<string>;
+  /** Track which run elements have already been wrapped. */
+  wrappedRuns: Set<Element>;
+  /** Source bookmark markers cloned into inserted deleted/moveFrom content. */
+  emittedSourceBookmarkMarkers: Set<Element>;
+  /** Trailing marker clones emitted for a source run. */
+  trailingSourceBookmarkClones: Map<Element, Element[]>;
+}
+
+export function createRevisionIdState(preservedRoots: readonly Element[] = []): RevisionIdState {
+  const state: RevisionIdState = {
+    nextId: 1,
+    usedRevisionIds: new Set(),
+    moveRangeIds: new Map(),
+    generatedMoveRangeMarkers: new Set(),
+    generatedFormatChangeIds: new Set(),
+    wrappedRuns: new Set(),
+    emittedSourceBookmarkMarkers: new Set(),
+    trailingSourceBookmarkClones: new Map(),
+  };
+  for (const root of preservedRoots) seedRevisionIdsFromMarkup(state, root);
+  return state;
+}
+
+function canonicalDecimalId(value: string): string | null {
+  const collapsed = value.trim();
+  if (!/^[+-]?\d+$/.test(collapsed)) return null;
+  try {
+    return BigInt(collapsed).toString();
+  } catch {
+    return null;
+  }
+}
+
+/**
+ * Reserve every numeric `w:id` in markup that can survive into the comparison
+ * output. Numeric aliases such as `1`, `01`, and `+1` reserve one identity.
+ *
+ * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.23
+ * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.24
+ * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.27
+ * @conformance ECMA-376 edition 5, Part 1 § 17.13.5.28
+ * @see https://github.com/UseJunior/safe-docx/issues/446
+ */
+export function seedRevisionIdsFromMarkup(state: RevisionIdState, root: Element): void {
+  const visit = (element: Element): void => {
+    const rawId = element.getAttribute('w:id');
+    if (rawId !== null) {
+      const canonical = canonicalDecimalId(rawId);
+      if (canonical !== null) state.usedRevisionIds.add(canonical);
+    }
+    for (const child of childElements(element)) visit(child);
+  };
+  visit(root);
+}
+
+export function allocateRevisionId(state: RevisionIdState): number {
+  while (state.usedRevisionIds.has(String(state.nextId))) state.nextId++;
+  const id = state.nextId++;
+  state.usedRevisionIds.add(String(id));
+  return id;
+}
+
+/** Format a date for OOXML (ISO 8601 without milliseconds). */
+export function formatDate(date: Date): string {
+  return date.toISOString().replace(/\.\d{3}Z$/, 'Z');
+}
+
+function convertToDeletedText(element: Element): void {
+  if (element.tagName === 'w:t' || element.tagName === 'w:instrText') {
+    const replacement = createRevisionElement(
+      element.tagName === 'w:t' ? 'w:delText' : 'w:delInstrText',
+    );
+    while (element.firstChild) replacement.appendChild(element.firstChild);
+    for (let index = 0; index < element.attributes.length; index++) {
+      const attribute = element.attributes[index]!;
+      replacement.setAttribute(attribute.name, attribute.value);
+    }
+    element.parentNode?.replaceChild(replacement, element);
+    return;
+  }
+  for (const child of childElements(element)) convertToDeletedText(child);
+}
+
+export type TrackChangeTag = 'w:ins' | 'w:del' | 'w:moveFrom' | 'w:moveTo';
+
+export interface WrapRunOptions {
+  run: Element;
+  tagName: TrackChangeTag;
+  author: string;
+  dateStr: string;
+  state: RevisionIdState;
+  convertTextToDelText?: boolean;
+}
+
+export function wrapRunWithTrackChange(options: WrapRunOptions): boolean {
+  const {
+    run,
+    tagName,
+    author,
+    dateStr,
+    state,
+    convertTextToDelText = false,
+  } = options;
+  if (state.wrappedRuns.has(run) || !run.parentNode) return false;
+
+  if (convertTextToDelText) convertToDeletedText(run);
+  const wrapper = createRevisionElement(tagName, {
+    'w:id': String(allocateRevisionId(state)),
+    'w:author': author,
+    'w:date': dateStr,
+  });
+  wrapElement(run, wrapper);
+  state.wrappedRuns.add(run);
+  return true;
+}
+
+export function addParagraphMarkRevisionMarker(
+  paragraph: Element,
+  markerTag: 'w:ins' | 'w:del',
+  author: string,
+  dateStr: string,
+  state: RevisionIdState,
+): void {
+  let paragraphProperties = findChildByTagName(paragraph, 'w:pPr');
+  if (!paragraphProperties) {
+    paragraphProperties = createRevisionElement('w:pPr');
+    paragraph.insertBefore(paragraphProperties, paragraph.firstChild);
+  }
+
+  const existingMarker = findParagraphMarkRevisionMarker(paragraphProperties, markerTag);
+  let runProperties = findChildByTagName(paragraphProperties, 'w:rPr');
+  if (!runProperties) {
+    runProperties = createRevisionElement('w:rPr');
+    const insertBefore =
+      findChildByTagName(paragraphProperties, 'w:sectPr') ??
+      findChildByTagName(paragraphProperties, 'w:pPrChange') ??
+      null;
+    paragraphProperties.insertBefore(runProperties, insertBefore);
+  }
+
+  if (existingMarker) {
+    existingMarker.setAttribute('w:author', author);
+    existingMarker.setAttribute('w:date', dateStr);
+    if (!existingMarker.getAttribute('w:id')) {
+      existingMarker.setAttribute('w:id', String(allocateRevisionId(state)));
+    }
+    placeParagraphMarkRevisionMarker(runProperties, existingMarker, markerTag);
+    return;
+  }
+
+  const marker = createRevisionElement(markerTag, {
+    'w:id': String(allocateRevisionId(state)),
+    'w:author': author,
+    'w:date': dateStr,
+  });
+  placeParagraphMarkRevisionMarker(runProperties, marker, markerTag);
+}
+
+export function placeParagraphMarkRevisionMarker(
+  runProperties: Element,
+  marker: Element,
+  markerTag: 'w:ins' | 'w:del',
+): void {
+  const insertionSibling =
+    markerTag === 'w:del' ? findChildByTagName(runProperties, 'w:ins') : null;
+  if (insertionSibling) {
+    if (insertionSibling.nextSibling !== marker) insertAfterElement(insertionSibling, marker);
+  } else if (runProperties.firstChild !== marker) {
+    runProperties.insertBefore(marker, runProperties.firstChild);
+  }
+}
+
+export function findParagraphMarkRevisionMarker(
+  paragraphProperties: Element,
+  markerTag: 'w:ins' | 'w:del',
+): Element | null {
+  for (const child of childElements(paragraphProperties)) {
+    if (child.tagName !== 'w:rPr') continue;
+    const marker = findChildByTagName(child, markerTag);
+    if (marker) return marker;
+  }
+  return null;
+}
